refactor code to modules
